corrections, lint and refactor
